refactor(adr-061): split service.py's dump orchestration into a thin facade
Splits abicheck/service.py's native-binary dump orchestration
(run_dump/_run_dump_uncached/_finish_native_snapshot/
_apply_native_provenance/_emit/_dump_elf/_extract_pdb_debug/_dump_pe/
_dump_macho, plus the _HEADER_GRAPH_ENABLED/_HEADER_GRAPH_INCLUDES_ENABLED
constants) out into two new sibling leaf modules -- service_dump_native.py
(shared orchestration + ELF tail) and service_dump_native_pe.py (PE/Mach-O
tail, split into its own file to stay under the 800-line production cap a
brand-new file gets no debt-ledger baseline to grow into) -- following the
same "extract to a service_<name> leaf, re-export verbatim" pattern
service_metadata_attach/service_header_graph_attach/service_header_scoped/
service_render/service_scan/service_compare_pipeline/service_dump_pipeline
already established.

service.py drops from 1763 to 872 lines. Every public and private name
involved is re-exported unchanged (explicit `as` self-aliases where mypy's
no_implicit_reexport requires it). architecture/modules.yaml registers both
new files in frozen_root_families.service_ and legacy_root_modules;
architecture/debt.yaml's service.py entry's baseline_lines is lowered to
match. scripts/check_ai_readiness.py's IMPORT_CYCLE_ALLOWLIST gains
service_dump_native on the same "split of an existing member" terms
service_header_graph_attach/service_compare_pipeline/service_dump_pipeline
were each signed off under before it.

Fixes ~65 existing test-patch call sites (tests/test_service_unit.py,
tests/test_service_clang_layout_renumbering.py,
tests/test_lambda_identity_ordinal.py, tests/test_pdb_provenance.py) whose
patch targets no longer intercept the call they were written to observe,
now that the call resolves against the new modules' own globals instead of
service.py's -- the same test-patch gotcha workflows/extraction.py's own
docstring documents. See the changelog fragment for the full accounting of
what was found and fixed, including a second review pass that caught
multi-line patch() calls and a dropped _try_attach_* re-export the first
pass missed.

Only this one block moved; resolve_input/sniff_text_format/
compare_snapshots/policy-loading and the existing tail-of-file re-export
blocks stay in service.py -- see the PR body for the investigation and
what's left as follow-up work.

Verification: check_architecture.py 0 errors, check_ai_readiness.py 0
errors (152 pre-existing warnings, unchanged), mypy abicheck/ clean (17
pre-existing yaml-stub baseline errors only), ruff check clean, full fast
unit suite 32969 passed / 129 skipped / 4 xfailed / 1 failed (the one
failure, test_dumper_manifest.py's TU-ordering test, is a pre-existing
system-load-dependent flake unrelated to this change -- reproduces
identically on main and passes in isolation on both main and this branch).
