+++ abicheck/service_dump_native_pe.py
@@ -1,0 +1,275 @@
+# Copyright 2026 Nikolay Petrov
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""PE and Mach-O native-binary dump: the ``_dump_pe``/``_dump_macho`` tail of
+``service_dump_native._run_dump_uncached``.
+
+Split out of ``service_dump_native.py`` for the identical AI-readiness
+file-size reason that module's own docstring gives for splitting out of
+``service.py`` in the first place — a genuinely new file has no debt-ledger
+baseline to grow into (``scripts/check_architecture.py``'s ``new-file-size``
+check), so the ELF/PE-Mach-O split had to land as two files rather than one
+just over the 800-line production cap. ``service.py`` re-exports both names
+below, so ``from abicheck.service import _dump_pe`` keeps resolving
+unchanged.
+
+**Test-patch note (see ``service_dump_native.py``'s own, fuller version of
+this same note):** a test substituting ``_dump_pe``/``_dump_macho`` for a
+call made *inside* ``service_dump_native.py`` (i.e. from ``run_dump``) must
+patch ``abicheck.service_dump_native_pe.<name>`` — that is where the call
+this module supplies actually resolves from the *caller's* side, since
+``service_dump_native.py`` imports these two names directly from here.
+"""
+
+from __future__ import annotations
+
+import importlib as _importlib
+import logging
+from pathlib import Path
+from typing import TYPE_CHECKING
+
+from .errors import SnapshotError, ValidationError
+from .model import AbiSnapshot, EnumType, Function, RecordType, Visibility
+
+if TYPE_CHECKING:
+    from collections.abc import Callable
+
+    from .compile_context import CompileContext
+    from .dwarf_advanced import AdvancedDwarfMetadata
+    from .dwarf_metadata import DwarfMetadata
+
+# Deliberately the *parent* module's logger name -- same convention
+# ``service_dump_native.py``/``service_metadata_attach.py`` already use, so
+# a test capturing "abicheck.service" logging keeps working regardless of
+# which leaf module actually logs.
+_logger = logging.getLogger("abicheck.service")
+
+# Bound lazily rather than imported statically to avoid re-creating the
+# import-cycle hazard ``service_dump_native.py``'s own
+# ``_try_header_scoped_dump`` binding documents (``service_header_scoped``
+# reaches ``service_scan``, which reaches back through the pre-existing,
+# already-baselined cli_buildsource/scan_engine SCC).
+_service_header_scoped = _importlib.import_module(".service_header_scoped", __package__)
+# Explicitly typed (not left as the `Any` importlib.import_module's attribute
+# access would otherwise infer) so a caller returning this call's result
+# still gets a real return-type check instead of a silent `no-any-return`.
+_try_header_scoped_dump: Callable[..., tuple[AbiSnapshot | None, str | None]] = (
+    _service_header_scoped._try_header_scoped_dump
+)
+del _service_header_scoped
+
+
+def _extract_pdb_debug(
+    path: Path, pdb_path: Path | None
+) -> tuple[DwarfMetadata | None, AdvancedDwarfMetadata | None]:
+    """Locate and parse a PDB for *path*.
+
+    Returns ``(dwarf_meta, dwarf_adv)`` or ``(None, None)`` when no PDB is found
+    or parsing fails.  PDB extraction is best-effort and never fatal.
+    """
+    try:
+        from .pdb_metadata import parse_pdb_debug_info
+        from .pdb_utils import locate_pdb
+
+        pdb_file = locate_pdb(path, pdb_path_override=pdb_path, allow_network=False)
+        if pdb_file is not None:
+            meta, adv = parse_pdb_debug_info(pdb_file)
+            _logger.info("PDB debug info loaded from %s", pdb_file)
+            return meta, adv
+        _logger.debug("No PDB file found for %s", path)
+    except Exception as exc:  # noqa: BLE001
+        _logger.warning("PDB parsing failed for %s: %s", path, exc)
+    return None, None
+
+
+def _dump_pe(
+    path: Path,
+    version: str,
+    *,
+    headers: list[Path] | None = None,
+    includes: list[Path] | None = None,
+    lang: str = "c++",
+    lang_explicit: bool = False,
+    pdb_path: Path | None = None,
+    header_backend: str = "auto",
+    compile: CompileContext | None = None,
+    public_headers: list[Path] | None = None,
+    public_header_dirs: list[Path] | None = None,
+    include_labels: dict[Path, str] | None = None,
+) -> AbiSnapshot:
+    """Dump a PE binary (Windows DLL) to an ABI snapshot.
+
+    When *headers* are supplied the ABI surface is scoped to declarations in
+    those public headers via castxml (mirroring ``abidw --headers-dir``).  If
+    castxml is unavailable or no header declaration matches an exported symbol,
+    scoping is skipped (with a warning) and the full export table is used.
+    """
+    from .pe_metadata import parse_pe_metadata
+
+    try:
+        pe_meta = parse_pe_metadata(path)
+    except ImportError as exc:
+        raise SnapshotError(str(exc)) from exc
+    except (RuntimeError, OSError, ValueError) as exc:
+        raise SnapshotError(f"Failed to parse PE '{path}': {exc}") from exc
+
+    if not pe_meta.machine:
+        raise SnapshotError(
+            f"Failed to extract PE metadata from '{path}'. "
+            "The file may be corrupt or not a valid PE binary."
+        )
+    if not pe_meta.exports:
+        raise ValidationError(
+            f"PE file '{path}' has no exports (named or ordinal). "
+            "Verify the file is a valid DLL."
+        )
+
+    dwarf_meta, dwarf_adv = _extract_pdb_debug(path, pdb_path)
+
+    scope_fallback: str | None = None
+    if headers:
+        scoped, scope_fallback = _try_header_scoped_dump(
+            "pe",
+            path,
+            headers,
+            includes or [],
+            version,
+            lang,
+            lang_explicit=lang_explicit,
+            header_backend=header_backend,
+            compile=compile,
+            public_headers=public_headers,
+            public_header_dirs=public_header_dirs,
+            include_labels=include_labels,
+        )
+        if scoped is not None:
+            # Preserve any PDB debug info alongside the header-scoped surface.
+            if dwarf_meta is not None:
+                scoped.dwarf = dwarf_meta
+                scoped.dwarf_advanced = dwarf_adv
+            return scoped
+
+    funcs = [
+        Function(
+            name=(exp.name or f"ordinal:{exp.ordinal}"),
+            mangled=(exp.name or f"ordinal:{exp.ordinal}"),
+            return_type="?",
+            visibility=Visibility.PUBLIC,
+            is_extern_c=not (exp.name or "").startswith("?"),
+        )
+        for exp in pe_meta.exports
+    ]
+
+    # ADR-024 Phase 1 (PDB provenance): when header scoping was requested but
+    # castxml could not resolve a surface (commonly the MSVC C++-mangling
+    # gap), recover declared types -- with their defining source header --
+    # from PDB debug info, so public-header scoping still has a provenance
+    # signal to classify against. Bounded so default PE diffs are unaffected.
+    pdb_types: list[RecordType] = []
+    pdb_enums: list[EnumType] = []
+    if headers and dwarf_meta is not None:
+        from .pdb_model import model_types_from_dwarf_metadata
+
+        pdb_types, pdb_enums = model_types_from_dwarf_metadata(dwarf_meta)
+
+    return AbiSnapshot(
+        library=path.name,
+        version=version,
+        functions=funcs,
+        types=pdb_types,
+        enums=pdb_enums,
+        pe=pe_meta,
+        dwarf=dwarf_meta,
+        dwarf_advanced=dwarf_adv,
+        platform="pe",
+        scope_fallback=scope_fallback,
+    )
+
+
+def _dump_macho(
+    path: Path,
+    version: str,
+    *,
+    headers: list[Path] | None = None,
+    includes: list[Path] | None = None,
+    lang: str = "c++",
+    lang_explicit: bool = False,
+    header_backend: str = "auto",
+    compile: CompileContext | None = None,
+    public_headers: list[Path] | None = None,
+    public_header_dirs: list[Path] | None = None,
+    include_labels: dict[Path, str] | None = None,
+) -> AbiSnapshot:
+    """Dump a Mach-O binary (macOS dylib) to an ABI snapshot.
+
+    When *headers* are supplied the ABI surface is scoped to declarations in
+    those public headers via castxml; otherwise the full export table is used.
+    """
+    from .macho_metadata import parse_macho_metadata
+
+    try:
+        macho_meta = parse_macho_metadata(path)
+    except (RuntimeError, OSError, ValueError) as exc:
+        raise SnapshotError(f"Failed to parse Mach-O '{path}': {exc}") from exc
+
+    if (
+        not macho_meta.exports
+        and not macho_meta.install_name
+        and not macho_meta.dependent_libs
+    ):
+        raise SnapshotError(
+            f"Mach-O file '{path}' has no exports or load-command metadata. "
+            "Verify the file is a valid dynamic library."
+        )
+
+    scope_fallback: str | None = None
+    if headers:
+        scoped, scope_fallback = _try_header_scoped_dump(
+            "macho",
+            path,
+            headers,
+            includes or [],
+            version,
+            lang,
+            lang_explicit=lang_explicit,
+            header_backend=header_backend,
+            compile=compile,
+            public_headers=public_headers,
+            public_header_dirs=public_header_dirs,
+            include_labels=include_labels,
+        )
+        if scoped is not None:
+            return scoped
+
+    funcs = [
+        Function(
+            name=exp.name,
+            mangled=exp.name,
+            return_type="?",
+            visibility=Visibility.PUBLIC,
+            is_extern_c=not exp.name.startswith("_Z"),
+        )
+        for exp in macho_meta.exports
+        if exp.name
+    ]
+    return AbiSnapshot(
+        library=path.name,
+        version=version,
+        functions=funcs,
+        macho=macho_meta,
+        platform="macho",
+        scope_fallback=scope_fallback,
+    )
+
