+++ architecture/modules.yaml
@@ -289,6 +289,8 @@
       "service_compare_evidence.py",
       "service_compare_pipeline.py",
       "service_dump_cache.py",
+      "service_dump_native.py",
+      "service_dump_native_pe.py",
       "service_dump_pipeline.py",
       "service_header_graph_attach.py",
       "service_header_scoped.py",
@@ -664,6 +666,8 @@
     "service_compare_evidence.py",
     "service_compare_pipeline.py",
     "service_dump_cache.py",
+    "service_dump_native.py",
+    "service_dump_native_pe.py",
     "service_dump_pipeline.py",
     "service_header_graph_attach.py",
     "service_header_scoped.py",
